# authApi/src/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1.0.0",
    "title": "OsanScheduler Authentication API",
    "description": "OsanScheduler provides team/site scheduling information"
  },
  "host": "localhost:7004",
  "basePath": "/api/authentication",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "authorization": {
      "type": "apiKey",
      "name": "authorization",
      "in": "header",
      "description": "Authentication token"
    },
    "refreshtoken": {
      "type": "apiKey",
      "name": "refreshtoken",
      "in": "header",
      "description": "Refresh token"
    }
  },
  "paths": {
    "/user/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "authorization": [],
            "refreshtoken": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "authorization": [],
            "refreshtoken": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/users": {
      "get": {
        "description": "",
        "security": [
          {
            "authorization": [],
            "refreshtoken": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/user": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "emailAddress": {
                  "example": "any"
                },
                "firstName": {
                  "example": "any"
                },
                "middleName": {
                  "example": "any"
                },
                "lastName": {
                  "example": "any"
                },
                "permissions": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "security": [
          {
            "authorization": [],
            "refreshtoken": []
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "any"
                },
                "field": {
                  "example": "any"
                },
                "value": {
                  "example": "any"
                },
                "subid": {
                  "example": "any"
                },
                "subfield": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "security": [
          {
            "authorization": [],
            "refreshtoken": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}